# docs/openapi.yaml
openapi: 3.0.3
info:
  title: АИС «Реестр обязательных требований» — HTTP API
  version: "1.0.0"
  description: |
    Спецификация фактического API веб-приложения (Next.js App Router, 46 маршрутов).
    Составлена по коду `src/app/api/**` (2026-07-02), см. docs/architecture (комплект ГОСТ 34).

    **Аутентификация:** JWT в httpOnly-cookie `auth_token` (выдаётся `/api/auth/login`, срок 7 суток).
    **Роли:** `admin` (супер-админ МНЭ) · `mne` (сотрудник МНЭ: всё как admin, кроме выгрузки каталога и создания учёток уровня МНЭ) · `moderator` (орган + поддерево организаций) ·
    `expert` (назначенные сферы ∩ органы). Скоупы возвращает `/api/auth/me`.
    **Ошибки:** единый формат `{ "error": string }`; 401 — не авторизован, 403 — нет прав/чужой орган,
    404 — не найдено, 429 — превышен лимит (голосование), 5xx — внутренняя/внешний сервис.

    Раздел «Легаси-экспертиза» обслуживает исторический MVP-контур (requirement_cards/голосования)
    и не развивается — см. документ 09 плана (К5).
servers:
  - url: https://reestr-web-production.up.railway.app
    description: Продуктив (Railway)
  - url: http://localhost:3000
    description: Локальная разработка
security:
  - cookieAuth: []
...
paths:
...
  /api/admin/users/{id}/authorities:
    put:
      tags: [Администрирование]
      summary: Назначить органы (плоский скоуп)
      parameters: [ { name: id, in: path, required: true, schema: { type: integer } } ]
      requestBody: { content: { application/json: { schema: { type: object, properties: { authorities: { type: array, items: { type: string } } } } } } }
      responses: { "200": { description: OK } }
  /api/admin/users/{id}/orgs:
    put:
      tags: [Администрирование]
      summary: Назначить узлы иерархии (доступ = узел + потомки)
      parameters: [ { name: id, in: path, required: true, schema: { type: integer } } ]
      requestBody: { content: { application/json: { schema: { type: object, properties: { orgs: { type: array, items: { type: integer } } } } } } }
      responses: { "200": { description: OK } }
...
  /api/admin/agreement:
    get:
      tags: [Легаси-экспертиза]
      summary: Согласованность голосов v1
      parameters:
        - { name: sphere, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: iteration_id, in: query, schema: { type: integer } }
      responses: { "200": { description: OK } }
  /api/admin/gold-labels:
    get:
      tags: [Легаси-экспертиза]
      summary: Голд-метки v1
      parameters:
        - { name: sphere, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: iteration_id, in: query, schema: { type: integer } }
        - { name: min_votes, in: query, schema: { type: integer } }
        - { name: gold_confidence, in: query, schema: { type: number } }
      responses: { "200": { description: OK } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT (7 суток), выдаётся /api/auth/login
...